rename convertor code cleanup and documentation updates
rename convertor code cleanup and documentation updates 2
