+++ test/tests/meshgenerators/cut_mesh_by_plane_generator/tests
@@ -9,7 +9,6 @@
       exodiff = 'simple_hex_cut_in.e'
       cli_args = '--mesh-only simple_hex_cut_in.e 
                   Mesh/gmg/elem_type=HEX8'
-      allow_warnings = true
       mesh_mode = 'REPLICATED'
       recover = false
       detail ='cut a 3D mesh consisting of only HEX8 elements.'
@@ -20,7 +19,6 @@
       exodiff = 'simple_prism_cut_in.e'
       cli_args = '--mesh-only simple_prism_cut_in.e 
                   Mesh/gmg/elem_type=PRISM6'
-      allow_warnings = true
       mesh_mode = 'REPLICATED'
       recover = false
       detail ='cut a 3D mesh consisting of only PRISM6 elements.'
@@ -31,7 +29,6 @@
       exodiff = 'simple_pyramid_cut_in.e'
       cli_args = '--mesh-only simple_pyramid_cut_in.e 
                   Mesh/gmg/elem_type=PYRAMID5'
-      allow_warnings = true
       mesh_mode = 'REPLICATED'
       recover = false
       detail ='cut a 3D mesh consisting of only PYRAMID5 elements.'
@@ -42,7 +39,6 @@
       exodiff = 'simple_tet_cut_in.e'
       cli_args = '--mesh-only simple_tet_cut_in.e 
                   Mesh/gmg/elem_type=TET4'
-      allow_warnings = true
       mesh_mode = 'REPLICATED'
       recover = false
       detail ='cut a 3D mesh consisting of only TET4 elements.'
@@ -52,7 +48,6 @@
       input = 'hex_prism_cut.i'
       exodiff = 'hex_prism_cut_in.e'
       cli_args = '--mesh-only'
-      allow_warnings = true
       mesh_mode = 'REPLICATED'
       recover = false
       detail ='cut a 3D mesh with mixed element types.'
