+++ test/tests/meshgenerators/elements_to_tetrahedrons_convertor/gold/simple_pyramid_convert_in.e
