+++ test/tests/meshgenerators/elements_to_tetrahedrons_convertor/simple_convert.i
@@ -21,7 +21,7 @@
     block_id = 2
   []
   [convert]
-    type = TetrahedralElementsConvertor
+    type = ElementsToTetrahedronsConverter
     input = block_2
   []
 []
